- Added provider to EditQuiz.
- Separated widgets of edit_quiz_view.dart in different files.

- Added new subdirectory of view called edit_quiz.

- Fixed design of edit quiz view.
